Clean /usr/local Git repository before calling brew doctor

+++ docs/hudson/OMERO-homebrew-install.sh
@@ -7,11 +7,9 @@ set -x
 
 export ICE_VERSION=${ICE_VERSION:-zeroc-ice33}
 export OMERO_ALT=${OMERO_ALT:-ome/alt}
-export BREW_DIR=${BREW_DIR:-/usr/local}
-export VENV_DIR=${VENV_DIR:-$BREW_DIR/virtualenv}
+export VENV_DIR=${VENV_DIR:-/usr/local/virtualenv}
 export PSQL_DIR=${PSQL_DIR:-/usr/local/var/postgres}
 export OMERO_DATA_DIR=${OMERO_DATA_DIR:-/tmp/var/OMERO.data}
-export JOB_WS=`pwd`
 export BREW_OPTS=${BREW_OPTS:-}
 export SCRIPT_NAME=${SCRIPT_NAME:-OMERO.sql}
 VENV_URL=${VENV_URL:-https://raw.github.com/pypa/virtualenv/master/virtualenv.py}
@@ -22,56 +20,74 @@ else
     CURL="curl ${CURL_OPTS-} -O"
 fi
 
-###################################################################
-# Homebrew & pip uninstallation
-###################################################################
-
-# Remove existing PIP virtualenvironment
-if [ -f "$VENV_DIR/bin/pip" ]
-then
-    echo "Removing pip-installed packages IN $VENV_DIR"
-    # Solve Cython uninstallation error exit
-    ($VENV_DIR/bin/pip freeze -l | grep Cython && $VENV_DIR/bin/pip freeze uninstall -y Cython) || echo "Cython uninstalled"
-
-    for plugin in $($VENV_DIR/bin/pip freeze -l); do
-        packagename=$(echo "$plugin" | awk -F == '{print $1}')
-        echo "Uninstalling $packagename..."
-        $VENV_DIR/bin/pip uninstall -y $packagename
-    done
-
-    echo "Deleting $VENV_DIR"
-    rm -rf $VENV_DIR
+# Test whether this script is run in a job environment
+JOB_NAME=${JOB_NAME:-}
+if [[ -n $JOB_NAME ]]; then
+    DEFAULT_TESTING_MODE=true
+else
+    DEFAULT_TESTING_MODE=false
 fi
+TESTING_MODE=${TESTING_MODE:-$DEFAULT_TESTING_MODE}
 
-# Remove Homebrew installation
-if [ -d "$BREW_DIR/.git" ]
-then
-    echo "Uninstalling Homebrew"
-    rm -rf $BREW_DIR/Cellar $BREW_DIR/.git && $BREW_DIR/bin/brew cleanup
+if [ $TESTING_MODE ]; then
+    ###################################################################
+    # Homebrew & pip uninstallation
+    ###################################################################
+
+    # Remove existing PIP virtualenvironment
+    if [ -f "$VENV_DIR/bin/pip" ]
+    then
+        echo "Removing pip-installed packages IN $VENV_DIR"
+        # Solve Cython uninstallation error exit
+        ($VENV_DIR/bin/pip freeze -l | grep Cython && $VENV_DIR/bin/pip freeze uninstall -y Cython) || echo "Cython uninstalled"
+
+        for plugin in $($VENV_DIR/bin/pip freeze -l); do
+            packagename=$(echo "$plugin" | awk -F == '{print $1}')
+            echo "Uninstalling $packagename..."
+            $VENV_DIR/bin/pip uninstall -y $packagename
+        done
+
+        echo "Deleting $VENV_DIR"
+        rm -rf $VENV_DIR
+    fi
+
+    # Remove Homebrew installation
+    if [ -d "/usr/local/.git" ]
+    then
+        echo "Uninstalling Homebrew"
+        rm -rf /usr/local/Cellar && /usr/local/bin/brew prune
+        rm -rf /usr/local/.git && /usr/local/bin/brew cleanup
+    fi
+
+    if [ -d "/usr/local/Library/Taps" ]
+    then
+        echo "Cleaning Homebrew taps"
+        rm -rf /usr/local/Library/Taps
+    fi
 fi
-
-if [ -d "$BREW_DIR/Library/Taps" ]
-then
-    echo "Cleaning Homebrew taps"
-    rm -rf $BREW_DIR/Library/Taps
-fi
-
 ###################################################################
 # Homebrew installation
 ###################################################################
 
 # Install Homebrew in /usr/local
 ruby -e "$(curl -fsSL https://raw.github.com/mxcl/homebrew/go)"
-cd $BREW_DIR
-
-# Clean cache before any operation to test full installation
-rm -rf $(bin/brew --cache)
+cd /usr/local
 
 # Install git if not already installed
 bin/brew list | grep "\bgit\b" || bin/brew install git
 
-# Update homebrew and run brew doctor
+# Update Homebrew
 bin/brew update
+
+if [ $TESTING_MODE ]; then
+    # Clean cache before any operation to test full installation
+    rm -rf $(bin/brew --cache)
+
+    # Clean the Git repository
+    git clean -df
+fi
+
+# Run brew doctor
 export PATH=$(bin/brew --prefix)/bin:$PATH
 bin/brew doctor
 
@@ -89,16 +105,19 @@ set +u
 source $VENV_DIR/bin/activate
 set -u
 
-# Install scc tools
-$VENV_DIR/bin/pip install -U scc || echo "scc installed"
-
 # Tap ome-alt library
 bin/brew tap $OMERO_ALT || echo "Already tapped"
 
-# Merge homebrew-alt PRs
-cd Library/Taps/${OMERO_ALT/\//-}
-$VENV_DIR/bin/scc merge master
-cd $BREW_DIR
+if [ $TESTING_MODE ]; then
+    # Install scc tools
+    $VENV_DIR/bin/pip install -U scc || echo "scc installed"
+
+    # Merge homebrew-alt PRs
+    cd Library/Taps/${OMERO_ALT/\//-}
+    $VENV_DIR/bin/scc merge master
+fi
+
+cd /usr/local
 
 ###################################################################
 # Bio-Formats installation
@@ -117,7 +136,7 @@ bin/brew install omero $BREW_OPTS
 bin/brew install postgres
 
 # Install additional Python dependencies
-bash $JOB_WS/docs/install/python_deps.sh
+bash bin/omero_python_deps
 
 # Set environment variables
 export ICE_CONFIG=$(bin/brew --prefix omero)/etc/ice.config
